fix mpi issue where send_problem can be null

+++ lib/common/configuration.h
@@ -43,7 +43,7 @@ class configuration {
     int random_k = 0;
     size_t bfs_size = 0;
     size_t threads = 1;
-    size_t preset_percentage = 0;
+    double preset_percentage = 0;
     size_t total_terminals = 0;
     size_t print_cc = 0;
     bool disable_cpu_affinity = false;
